// src/models/office/LSS_V_LSSIRA05.model.ts
export interface LSS_V_LSSIRA05 {
  id: any;
  TRANS_DATE: Date;
  YEAR: string;
  MONTH: string;
  QUARTER: string;
  LOGIN_NAME: string;
  SYSCODE: string;
  ACCOUNT_TYPE_NAME: string;
  ACCOUNT_TYPE_CODE: string;
  IP_ADDRESS: string;
  SYSTEM_CODE: string;
  MENU_CODE: string;
  MENU_NAME: string;
  FUNCTION_NAME: string;
  FUNCTION_CODE: string;
  SECTION_CODE: string;
  SECTION_NAME: string;
  POSITION_CODE: string;
  POSITION_NAME: string;
  ACCOUNT_TYPE: string;
  QTY: number;
}